+++ react/components/Modal/index.js
@@ -1,8 +1,16 @@
 import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
 import ResponsiveModal from 'react-responsive-modal'
+import Button from '../Button'
 
 class Modal extends PureComponent {
+  handleCancelation = () => {
+    this.props.cancelation && this.props.cancelation()
+  }
+  handleConfirmation = () => {
+    this.props.confirmation && this.props.confirmation()
+  }
+
   render() {
     const {
       isOpen,
@@ -10,6 +18,10 @@ class Modal extends PureComponent {
       onClose,
       closeOnEsc,
       closeOnOverlayClick,
+      confirmation,
+      confirmationLabel,
+      cancelation,
+      cancelationLabel,
       showCloseIcon,
     } = this.props
 
@@ -43,6 +55,29 @@ class Modal extends PureComponent {
         closeIconSize={18}
       >
         {this.props.children}
+
+        {confirmation && (
+          <div className="vtex-modal__confirmation flex justify-end mt8">
+            {cancelation && (
+              <span className="mr4">
+                <Button
+                  size="small"
+                  variation="secondary"
+                  onClick={this.handleCancelation}
+                >
+                  {cancelationLabel}
+                </Button>
+              </span>
+            )}
+            <Button
+              size="small"
+              variation="primary"
+              onClick={this.handleConfirmation}
+            >
+              {confirmationLabel}
+            </Button>
+          </div>
+        )}
       </ResponsiveModal>
     )
   }
@@ -71,6 +106,12 @@ Modal.propTypes = {
   closeOnOverlayClick: PropTypes.bool,
   /** Show the close icon on upper right corner (default true) */
   showCloseIcon: PropTypes.bool,
+
+  /** Confirmation & Cancelation actions */
+  confirmation: PropTypes.func,
+  confirmationLabel: PropTypes.string,
+  cancelation: PropTypes.func,
+  cancelationLabel: PropTypes.string,
 }
 
 export default Modal